cm3628: adding doubleshot only Light level multiplication.
As currently used blob libsensors for doubleshot uses a high
divisor for determining the light level, we use a workaround
of light level multiplication in the kernel driver.
Only applied for CONFIG_MACH_DOUBLESHOT=y so it won't hurt
other devices libsensor interpretations.

Current Doubleshot AKM sensor sources segfaults for some
reason in JB (while worked in ICS), and thus for now we
add this workaround till we find a way to fix it in userspace.

Change-Id: I8d2f00fe0ffb2424865553b1fbe557010100cbf4