Update patch set 1
Patch Set 1:

What type of aic does ruby have? Is it different from pyramid/doubleshot

The issue that this fixes on doubleshot is that when phone mode and media is connected to headset it didn't play music, and couldn't place a call either with or without media. Also with this version when Bluetooth is on active or passive it deepsleeps well even when WiFi is on.

Patch-set: 1